+++ src/TheHolyCompiler-core/util/utils.h
@@ -35,7 +35,7 @@ class Utils {
 public:
 	static void CopyString(char*& dst, const char* const src);
 	static String GetPathFromFile(const String& filepath);
-	static unsigned int FindLine(const String& string, size_t index);
+	static uint32 FindLine(const String& string, uint64 index);
 	
 	static String ReadFile(const String& filename);
